refactor(sdk): limit conditional state to filesystem
