+++ libs/deepagents/tests/unit_tests/middleware/test_compact_tool.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from inspect import Parameter, signature
-from typing import Any, get_type_hints
+from typing import Any
 from unittest.mock import MagicMock, patch
 
 import pytest
@@ -14,7 +14,6 @@ from langgraph.types import Command
 from deepagents.backends.state import StateBackend
 from deepagents.middleware.summarization import (
     SummarizationMiddleware,
-    SummarizationState,
     SummarizationToolMiddleware,
     create_summarization_tool_middleware,
 )
@@ -720,20 +719,6 @@ def test_create_summarization_tool_middleware_returns_instance() -> None:
 
     assert isinstance(mw, SummarizationToolMiddleware)
     assert mw.tools[0].name == "compact_conversation"
-
-
-def test_create_summarization_tool_middleware_adds_state_backend_files() -> None:
-    model = GenericFakeChatModel(messages=iter([AIMessage(content="ok")]))
-    model.profile = {"max_input_tokens": 120_000}
-    middleware = create_summarization_tool_middleware(model, StateBackend())
-
-    assert "files" in get_type_hints(middleware.state_schema, include_extras=True)
-
-
-def test_summarization_tool_middleware_ignores_invalid_state_schema() -> None:
-    middleware = SummarizationToolMiddleware(MagicMock())
-
-    assert middleware.state_schema is SummarizationState
 
 
 def test_create_summarization_tool_middleware_accepts_system_prompt() -> None:
